fix(slack): tighten final installer polish
